requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-identity
azure-mgmt-resource
azure-mgmt-cosmosdb
azure-mgmt-storage

azure-cosmos
azure-storage-blob
azure-functions

PyGithub
openai ==0.28